Make ContentPicker work with NestedSharedElements
Prior to ag/30638736 we did not have a way to track a parent transition
from within a NestedStl, therefore we could not animate an element based
on a transition of an ancestor. This meant that for sharedElements
across nested STLs we always had to render the element in the highest
ancestor.

With this change we can render sharedElements in any content and
therefore we can bring back and simplify the code to use ContentPickers
as before.

This change needs to make adjustments to how we calculate zIndex for
each content, otherwise for each nestedSTL the zIndex is counted from
0f, therefore the default content picker (HighestZOrder) has mostly
arbitrary results.

Bug: 376659778
Test: Added new unit tests
Flag: com.android.systemui.scene_container

+++ packages/SystemUI/compose/features/src/com/android/systemui/media/controls/ui/composable/MediaContentPicker.kt
@@ -41,8 +41,8 @@ object MediaContentPicker : StaticElementContentPicker {
     override fun contentDuringTransition(
         element: ElementKey,
         transition: TransitionState.Transition,
-        fromContentZIndex: Float,
-        toContentZIndex: Float,
+        fromContentZIndex: Long,
+        toContentZIndex: Long,
     ): ContentKey {
         return when {
             transition.isTransitioningBetween(Scenes.Lockscreen, Scenes.Shade) -> {
